requirements.txt
flask==2.1.1
pymysql==1.0.2
sqlalchemy==1.4.0
pyyaml==6.0
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
azure-core==1.1.1
azure-cosmos==4.1.0